% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ind_ids.R
\name{create_ind_ids}
\alias{create_ind_ids}
\title{new ID spec for 2nd round of CMB:
"year entered survey" + "\emph{" + "3 serial #" + "}" + "census ID code"}
\usage{
create_ind_ids(data, census_id_name, year)
}
\description{
new ID spec for 2nd round of CMB:
"year entered survey" + "\emph{" + "3 serial #" + "}" + "census ID code"
}
