% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{last_non_na_col}
\alias{last_non_na_col}
\title{Find the last non-NA column in each row of data frame}
\usage{
last_non_na_col(df)
}
\arguments{
\item{df}{A data.frame, tibble, etc.}
}
\value{
A vector of the names of the last non-NA column in each row.
}
\description{
Find the last non-NA column in each row of data frame
}
