% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recode_from_embedded}
\alias{recode_from_embedded}
\title{Recode a target column dynamically}
\usage{
recode_from_embedded(
  data,
  target_col,
  source_col_pattern,
  na_vals = c("997", "998", "999")
)
}
\arguments{
\item{data}{A data.frame, tibble, etc.}

\item{target_col}{The name of the column to be recoded.}

\item{source_col_pattern}{A regex to match the source columns to be used in
the recode.}

\item{na_vals}{Values to be ignored.}
}
\value{
None. This function is called for its side-effects only. The \code{target}
column is recoded.
}
\description{
This function recodes a target column dynamically using values from a range
of related columns
}
