% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ind_ids.R
\name{create_ind_ids}
\alias{create_ind_ids}
\title{Generate a vector of ID codes}
\usage{
create_ind_ids(data, census_id_name, year)
}
\arguments{
\item{data}{Input data with the \code{data.frame} type.}

\item{census_id_name}{The name of the column that holds grouping IDs. In the
CMB data, this is a census identifier, which are generated by Statistics
Canada.}

\item{year}{A string value representing the year to be included in the IDs.}
}
\value{
The input \code{data.frame} (passed to the \code{data} parameter) with a new
column holding the new individual IDs called \code{ind_id}.
}
\description{
\code{create_ind_ids} generates a new column of individual ID codes to track
individuals in panel survey data. The ID specification is a string in the
following format:
}
\details{
"year entered survey" + "\emph{" + "3 serial #" + "}" + "census ID code"
}
